Prevent negative trainer balances (#23)
* Add explicit trainer debt state

* Route trainer reversals into debt

* Expose trainer debt in private responses

* Migrate trainer debt balances

* Test trainer debt settlement

* Prevent deleting trainers with debt

+++ symfony/src/Trainer/Repository/TrainerRepository.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Trainer\Repository;
 
 use App\Trainer\Entity\Trainer;
+use App\Trainer\Exception\CannotDeleteTrainerException;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\DBAL\LockMode;
 use Doctrine\ORM\Exception\ORMException;
@@ -28,6 +29,12 @@ class TrainerRepository extends ServiceEntityRepository
 
     public function remove(Trainer $trainer): void
     {
+        if ($trainer->getDebt() !== 0) {
+            throw new CannotDeleteTrainerException(
+                'Cannot delete trainer while debt is not zero'
+            );
+        }
+
         $this->getEntityManager()->remove($trainer);
     }
